ai-engine/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
scikit-learn==1.3.2
pandas==1.5.3
numpy==1.26.2
soccerdata==1.3.1
